test: extract runtime hint fixtures
